updated requirements, refactored code so that two runners are present (classic and torch-based)
